feat: useEffect and useRef
React components are inserted into the DOM by React. So for functionality to requires manipulation of a component rendered in the DOM. It needs to be done only after the component is rendered and using complex javascript to access the rendered component. React provides a useRef hook that allows an object to be created that will hold a reference to any object for access. In the case of accessing DOM elements useRef can be used to provide the reference. The variable pointing to the useRef call can be placed as a `ref` prop to the desired component to be accessed. The useEffect hook occurs after the component is render allows the ref variable to provide an `.current` property the provides access to the underlying DOM element. The useEffect hook returns a function that is call during Unmount of component or changes to the dependencies of the useEffect hook. If resources need to be clean up the code to perform the operations should be done in the return function. An array which contains dependencies that requires the useEffect to be reran when changed can be provided