feat: useEffect for async fetch
useEffect is used for any side effects that needs to be performed by the component. It can be DOM related side effects, external storage side effects or API access side effects. One caveat of useEffect is that it expects the return value to be the callback function for cleanup. So the following code will not work. Since the async function returns a Promise

```
// this does not work, don't do this:
React.useEffect(async () => {
  const result = await doSomeAsyncThing()
  // do something with the result
})
```
In order to use async await in useEffect it should be wrapped in an inner function like the following:
```
React.useEffect(() => {
  async function effect() {
    const result = await doSomeAsyncThing()
    // do something with the result
  }
  effect()
})
```