feat: Error Boundary
Error handling can be done within the component code but any unhandled errors or runtime javascript errors will cause the entire React application to crash and render an empty page. React introduced the Error Boundary which must be a class component acts like a try-catch for React Components. It introduces a static getDerivedStateFromError method that will be called when an error occurs with an error object. This object can be set into the component state to render an alternative component listed as the FallbackComponent in the props. The ErrorBoundary will catch all errors occurring in the subtree it contains. A single ErrorBoundary can be placed for the entire application or multiple ErrorBoundaries can exist which will allow the trees in the application that did not fail to continue to work

Error boundaries do not catch errors for:
- Event handlers (learn more)
- Asynchronous code (e.g. setTimeout or requestAnimationFrame callbacks)
- Server side rendering
- Errors thrown in the error boundary itself (rather than its children)